[telemetry] refactor telemetry code
