+++ src/host/CMakeLists.txt
@@ -27,6 +27,7 @@
 #
 
 add_subdirectory(posix)
+add_subdirectory(telemetry)
 
 add_library(otbr-host
     async_task.cpp
@@ -37,8 +38,6 @@ add_library(otbr-host
     ncp_spinel.hpp
     rcp_host.cpp
     rcp_host.hpp
-    telemetry_retriever_border_agent.cpp
-    telemetry_retriever_border_agent.hpp
     thread_helper.cpp
     thread_helper.hpp
     thread_host.cpp
