+++ src/host/telemetry/telemetry_retriever_border_agent.hpp
@@ -26,6 +26,11 @@
  *    POSSIBILITY OF SUCH DAMAGE.
  */
 
+#ifndef OTBR_AGENT_TELEMETRY_RETRIEVER_BORDER_AGENT_HPP_
+#define OTBR_AGENT_TELEMETRY_RETRIEVER_BORDER_AGENT_HPP_
+
+#include "openthread-br/config.h"
+
 #if OTBR_ENABLE_TELEMETRY_DATA_API && OTBR_ENABLE_BORDER_AGENT
 
 #include <utility>
@@ -59,3 +64,5 @@ private:
 } // namespace otbr
 
 #endif // OTBR_ENABLE_TELEMETRY_DATA_API && OTBR_ENABLE_BORDER_AGENT
+
+#endif // OTBR_AGENT_TELEMETRY_RETRIEVER_BORDER_AGENT_HPP_
